Fix of Bug 678: Path navigation in main header is not displayed.
The navigation menu is realized using the chrome.js javascript library which scans the document for <a/>-Elements where attribute 'rel' is set. This attribute is interpreted as the identifier for the menu to display.

Pages using the old layout-system labeled the link leading to the start page '/' with 'rel=Start'. This is conform to the html-standard (see http://www.w3.org/TR/html401/types.html#type-links), but collides with chrome's misuse of the 'rel'-attribute.

As a bug fix, in every jsp-file each occurrence of

<h1 id="path"><a href="/" rel="Start">

is replaced by

<h1 id="path"><a href="/">