Quick-Fix for bug 682: 'Certain nested blocks in bibtex fields parsed wrong in bibtex display'
The error occurs when curly brackets are used within a non-standard field, e.g.
   error = { {abc}, {efg} }

This is valid for BibTeX. All non-standard fields are stored in a single string. We used regular expressions to parse such fields but due to regexp's limitations this can't succeed.

Using the bibtex parser seems to be too expensive - so we just print out misc field's unformatted content as is done in other views.