added new property to GenericParam: rawSearch. this property will be set with method setSearch and you can get it with getRawSearch. It contains search parameter like property search, but without added '+'s before every seasrch term.
with this and some changes in LuceneSearch*  now it is possible to search with AND, OR and NOT. Make sure that these words are UPPERCASE!

LuceneSearch* is also prepared to deliver results not by date order, but by relevance order.

With relevance order, default bibsonomy operator AND is set to OR and a document which contains more search terms than other is more relevant than the other. documents with same relevance will be ordered by date.