moved from InitUserFilter to UserDatabaseManager: "check, if user exists, if it is an ldap user and if it has to re-auth agains ldap server. if so, do it."
Now I get a NullPointerException in InitUserFilter, if it must re-auth to ldap. Therefore ldap re-auth should be done in UserDatabaseManager also. But if so, class Ldap has to move from webapp to somewhere, maybe database.

This commit is for documentation use only - nevertheless it is functional without use of ldap and almost functional with ldap