Fixed an error (at least i hope so!)
the "for each" js loop seemed to have problems on biblicious. ran well at home though!

now im not looping through the globalErrors array but only alerting the error at position 0. (which i hope is ok since there can only be one error in the array?!)

Uploading to test it!