enhancements for spring security:
1. when the user logs in with the main header login form he is redirected back to the page he used to log in
2. added an exception translation filter and an AccessDeniedNoticeException for redirecting the user to the login page with a notice message

(please find a better solution for the exception translation filter!)

Note: if you want to redirect a not logged in user to the login page just throw an AccessDeniedException (the one from spring!) or an AccessDeniedNoticeException if you want a notice param to be appended (e.g. /login?notice=login.post.bookmark). After the user logged in spring security redirects the user back to your controller